pyproject.toml
[tool.poetry]
name = "google-sheet-client"
version = "0.1.0"
description = "A lightweight Python-based Google Sheet Client to simplify the process of exchanging data with Google Sheets."
authors = ["Olivier Bénard"]
license = "MIT"
readme = "README.md"
packages = [ { include = "google_sheet_client" } ]

[tool.poetry.dependencies]
python = "^3.12"
black = "^25.1.0"
mypy = "^1.15.0"
pylint = "^3.3.4"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
python-dotenv = "^1.0.1"
gspread = "^6.1.4"
google-auth = "^2.38.0"
tenacity = "^9.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
